+++ src/network/scenario.rs
@@ -28,7 +28,7 @@ impl Scenario {
         for &key in self.b_tuples_fixed.keys().chain(global_waiting.keys()) {
             let scenario_draw = self.waiting_at(key);
             let global_draw = *global_waiting.get(&key).unwrap_or(&0);
-            let relative_draw = draw_fn.apply(global_draw as i32, scenario_draw as i32, self.slack);
+            let relative_draw = draw_fn.apply(global_draw as i64, scenario_draw as i64, self.slack);
             self.network_state.relative_draws.insert(key, relative_draw);
         }
     }
